requirements.txt
django==4.2.11
flake8==6.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-split-settings==1.3.0
django-cors-headers==4.3.1
uwsgi==2.0.25
django-bootstrap4==24.3
djangorestframework==3.15.1
drf-nested-routers==0.94.1